Separate useActiveExample hook from component
