+++ packages/oc-docs/src/components/Docs/Sidebar/Sidebar.tsx
@@ -12,8 +12,7 @@ import { useNavModel } from '../../../routing/hooks';
 import { normalizeSlug } from '../../../routing/resolve';
 import { OVERVIEW_SLUG, ENVIRONMENTS_SLUG } from '../../../routing/navModel';
 import { useDocsNavigate } from '../../../hooks';
-import { NavigationState } from 'src/hooks/useDocsNavigate';
-import { NavModel } from 'src/routing';
+import { useActiveExample } from './hook/useActiveExample';
 
 interface SidebarProps {
   onNavigate?: () => void;
@@ -131,37 +130,5 @@ const Sidebar: React.FC<SidebarProps> = ({ onNavigate, testId = 'sidebar' }) => 
     </StyledWrapper>
   );
 };
-
-export function useActiveExample(
-  model: NavModel,
-  activeSlug: string,
-  uuidToSlug: Map<string, string>,
-  onNavigate?: () => void
-) {
-  // The highlighted example rides on the navigation entry's state, not the URL,
-  // so it is not shareable/deep-linked, clears itself on any navigation, and is
-  // restored correctly by browser back/forward. It always belongs to the request
-  // currently shown.
-  const { state } = useLocation();
-  const docsNavigate = useDocsNavigate();
-  const exampleIndex = (state as NavigationState)?.exampleIndex;
-  const activeRequestUuid = getItemUuid(model.bySlug.get(activeSlug)?.item);
-  const activeExample =
-    exampleIndex != null && activeRequestUuid !== undefined
-      ? { requestUuid: activeRequestUuid, index: exampleIndex }
-      : null;
-
-  const goToExample = (requestUuid: string, index: number) => {
-    const slug = uuidToSlug.get(requestUuid);
-    if (slug == null) return;
-    docsNavigate(slug, { state: { exampleIndex: index } });
-    onNavigate?.();
-  };
-
-  return {
-    goToExample,
-    activeExample
-  };
-}
 
 export default Sidebar;
